pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Huy Phung <huypc2410@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "0.0.287"
openai = "^0.28.0"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
python-multipart = "^0.0.6"
pydantic = {extras = ["email"], version = "^2.4.1"}
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
